third_party/WebKit/LayoutTests/media/video-controls-focus-movement-on-hide.html

<!DOCTYPE html>
<title>Test that focus moves to "body" when controls fade out with a button focused.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-controls.js"></script>
<script src="media-file.js"></script>
<video controls></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");

    video.oncanplay = t.step_func(function() {
        assert_equals(document.activeElement, document.body);

        var playButton = mediaControlsButton(video, "play-button");
        playButton.focus();
        playButton.click();

        assert_equals(document.activeElement, video);

        runAfterHideMediaControlsTimerFired(t.step_func_done(function() {
            var controls = mediaControlsButton(video, "panel");
            assert_equals(getComputedStyle(controls).opacity, "0");
            assert_equals(document.activeElement, document.body);
        }), video);
    });

    video.src = findMediaFile("video", "content/test");
});
</script>
